Fix Tabs unregistering on re-render in frontend
<Tab /> components should mount and register themselves into <Tabs />
and only be unregistered when they are unmounted.
However the hook implementing such logic depends on the tab's title.
If a JSX component was passed inline as the title prop for a tab, title
would refer to a different React node at each render.
The fix involves accepting only strings to define a Tab's title, since
equal strings are referentially equal and do not force the tab to
unregister if the title doesn't actually change.

Signed-off-by: Davide Briani <davide.briani@ispirata.com>