Implement tenant-specific auth for APIs in frontend
Add the authentication form in the Login page to authenticate against a
specific tenant via tenantSlug + authToken.
Once authenticated, the backend endpoint is setup to hit the APIs
specific for the selected tenant.
The auth data is persisted via cookies for the session, or for a
year if the user selects `Keep me logged in` upon login.
The Login page accepts the URL query params `tenantSlug` and `authToken`
to perform automatic authentication via URL.

Signed-off-by: Davide Briani <davide.briani@ispirata.com>