Fix GoogleGeolocation discarding too much info before geolocating
The Google Geolocation provider is able to retrieve a position from the
list of WiFi scans for a device.
However only the scans with exactly the latest timestamp were kept and
used to compute the position, which resulted in discarding scans that
were probably performed together but that had a sub-second difference.
 This commit adopts `DateTime.diff` to compare timestamps and to keep
all those that differ less than a seconds from the latest one.

Signed-off-by: Davide Briani <davide.briani@ispirata.com>