Load and extract localized Device descriptions in order of preference
- Preload both the default locale of the tenant and the locale specified
by the user, if any.
- Extract the best match of translated descriptions according to locale
preferences when rendering System Models in the GraphQL layer.

This change ensures that the user is always presented with a translated
description for one of the preferred locales, including the tenant's
default, so that he can correctly see and manage descriptions at least
in the tenant's locale until a full i18n support is implemented.

Signed-off-by: Davide Briani <davide.briani@secomind.com>