Add namespace + key index to device_attributes table
Selectors will have to retrieve attributes using their namespace and key. Currently there's a
primary key (and unique index) on tenant_id + device_id + namespace + key, but
given that tenant_id is not passed in subquery, an additional index is needed to
efficiently retrieve attributes.
Also fix a test that was broken by this modification, making it independent from
the retrieved order of the attributes.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>