Refactor contexts surrounding Devices
Here's a summary of what this (big) commit is doing

- Separate `%Astarte.Device{}` from `%Devices.Device{}`
These two have different responsibilities: the former is responsible of
interacting with Astarte, creating the device in the database and updating it
with triggers, the latter is responsible of everything Edgehog builds upon the
Astarte device. The two map to the same database table so they share data
through it.

- Make the Astarte context accept an AppEngine client when doing its queries
instead of a device
This makes the Astarte context independent, with the Devices context calling
into it without circular dependencies

- Extract Capabilities to a separate context
Capabilities are not an Astarte concept, they're an Edgehog concept built upon
Astarte. The new structure reflects that, without polluting the Astarte context

- Move GraphQL modules around to better map to contexts
Create Geolocation and Capabilities types and resolvers

- Fix all tests to make them work with the new structure

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>