devices: explicitly preload Astarte resources in a function
Previously Astarte resources (realm and cluster) were preloaded in
appengine_client_from_device, but that meant that if the user requested 3
different Astarte fields, we would preload realm and cluster 3 times since they
got discarded every time.
On the other hand, we don't want to put realm and cluster preloads on devices by
default, since the realm contains the private key which is a quite big string
and we don't want to preload it on a device list without the need to use it.

This commit adds an explicit preload_astarte_resources_for_device/1 function
that takes care of preloading the Astarte realm and cluster. On the resolver
side, in queries we check if we're requesting any field that needs to access
Astarte, and if we do, we preload realm and cluster once before passing the
device to child fields. For mutations that need to query Astarte, we always
preload.

The commit also adds the explicit preload to tests that don't pass through
resolvers.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>