resolvers: always preload astarte resources for Edgehog.Device
The previous implementation, which looked at the projected resolution, was
actually a premature optimization and actually broke when the device was
requested through an interface (like the Node interface that was fixed in the
previous commit).
Remove this and always preload Astarte resources for devices. This kind of
optimization will be handled by Dataloader in the future anyway.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>