Allow different interfaces sets to support a capability
Previously, we were using a map to map a capability to the set of all interfaces
required to support it. This meant that a capability could be supported only
with a single interface set.
This commit changes that, allowing for a capability to be supported using
different interface sets. The first use case for this is OTA, since we already
support two different interface pairs for that, but this unlocks more general
developments where we change how we provide a specific capability while
maintaining backwards compatibility.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>