target: rename status pending to in_progress
This is more coherent with the naming in the rollout mechanism, which is
max_in_progress_updates. Pending also was ambiguous since it had the same name
than the pending OTA Operation, while here we represent the whole range of
states between the start of the rollout and the completion.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>