push_rollout: rename max_errors_percentage to max_failure_percentage
This fixes two different problems: one, the plural "errors" was grammatically
incorrect, and the correct concept is "failure", since that's the term we use
for exceptions which result in a final state, while error is a possibly
temporary condition.
Add a migration to rename the field in the database and replace all occurrences
both in frontend and backend. The only remaining occurrence is in the docs, but
they have to be completely revamped anyway so it doesn't make sense to touch
them here.

Close #287

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>