Add configuration file for Vite
Add the vite.config.ts file which specifies the build process.
- `server.open = true` is used to automatically open a browser's tab
when launching Vite's development environment. The flag is enabled to
respect previous habits, and the same goes for `server.port` to keep
using port 3000 for local development.
- `build.outDir` specifies the directory for the build artifacts. The
`build` directory is used so that the packaging pipeline don't need to
be modified, e.g. Dockerfile.
- `plugins` enhance the capabilities of Vite's build process,
integrating it with the Eslint linter, Relay's compiler and the
resolution rules of the existing tsconfig.json file.

Signed-off-by: Sergey Zakhlypa <sergey.zakhlypa@secomind.com>