fix(seeds): restore functionality outside docker
adds a dependency to load the .env file when it exists.

the method `read` is used to get back a map of the loaded values.
this is because the `load` method prioritizes variables in the .env file to system variables,
a behavior shared by all the other libraries taken in consideration
- dotenv_elixir
- dotenvy
- dotenv-parser

but we want a behavior consistent with docker, where exported variables are prioritized.

closes #393

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>