Add GraphQL mutation to request a forwarder session for a device
Add a `requestForwarderSession` mutation to allow clients to request a
device → Forwarder session for a specific deviceId.

To support the request, the backend does the following:
- It checks whether the device is online. If not, an error is returned.
- It queries the existing forwarder sessions for the device:
  - If there is a :connected session, the backend returns the details
about that session.
  - If there is a :connecting session, the backend returns the details
about that session.
  - If there is no available session, it generates a new session token
and sends a request to the device asking to establish that session. It
then returns the details about that session.

With this approach, any existing forwarder sessions is reused whenever
possible, and devices are not forced to establish a high number of
simultaneous connections.

Closes #445

Signed-off-by: Davide Briani <davide.briani@secomind.com>