Add button on the Device page to open a Remote Terminal session
When the device supports the REMOTE_TERMINAL capability, display a
button to request a forwarder session.
When the button is clicked, a forwarder session is requested and the
resulting credentials are used to open a new browser tab, navigating to
the URL of the cloud forwarder that exposes the TTYD service for the
device.
The button is disabled if the device is offline.

Signed-off-by: Davide Briani <davide.briani@secomind.com>