Forwarder.Session: porting to Ash
- Refactor the logic of fetching the forwarder sessions of a device as
an Ash calculation on the Device resource.
- Remove the `fetch_session/3` function from the ForwarderSession module
in the Astarte domain: we won't be needing it as a function in that
module since that logic can be extracted in an upper layer and it's not
a 1:1 mapping to Astarte APIs.
- Define a Forwarder.Session Ash resource to encapsulate the actions to
retrieve or request forwarder sessions. The resource exports a GraphQL
query that allows retrieving the status of a device's forwarder session
given its token, and a mutation that allows requesting an active session.
- Port the test relative to the previous GraphQL operations to perform
checks on the newly defined ones.

Signed-off-by: Davide Briani <davide.briani@secomind.com>