Geolocation tests: remove module stubs, prefer explicit mocking
This commit removes the stub implementations for geolocation and
geocoding modules.
The idea is that explicitly mocking the needed functions in test files
is usually preferred, as it makes it clear what assumptions are being
made during tests.

The whole GeolocationMockCase module is removed since, after removing
the `Mox.stub_with/2` directives, it only contains setup logic for Mox
which is not strictly related to geolocation.
Indeed, that setup logic is already performed by the GraphQLCase module,
or it could also be explicitly inserted in test files as needed.

Signed-off-by: Davide Briani <davide.briani@secomind.com>