RolloutMechanism: define Ash union containing PushRollout
Refactor the PushRollout embedded schema as an Ash embedded resource,
which is referenced by the newly introduced RolloutMechanism union type.

Besides porting the resource to Ash, the default values are changed for
some fields:
- `ota_request_retries` now defaults to 3 instead of 0.
- `ota_request_timeout_seconds` now defaults to 300 instead of 60.

This change should make the Executor logic better match the behaviour of
the Edgehog Device Runtime and the MQTT keepalive intervals.

Fixes #343

Signed-off-by: Davide Briani <davide.briani@secomind.com>