Make non-atomic validations run only when attributes are changing
Ash always runs all validations either on the changing value or on existing
data. This means that some validations couldn't be performed atomically, but we
are actually interested in validating those only when they're changing.
Add a where clause to those validations to only run them when the attribute
changes. This allows to make some actions atomic. The ones that are still not
atomic have been annotated to indicate why.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>