GraphQL: paginate "list" queries and relationships w/ Relay connections
Update all GraphQL query definitions by specifying that pagination and
Relay connections should be used.
The same approach is used to expose the relationships of resources.

At the same time, the default `read` action is preferred as the backing
bone of the functionality, remove the distinction of "get" and "list"
actions: this distinction only becomes relevant in code interfaces or
GraphQL APIs.

Close #522

Signed-off-by: Davide Briani <davide.briani@secomind.com>