Fix: add padding in Sidebar since Bootstrap does not apply it
Recent versions of Bootstrap were subject to refactoring to use CSS
variables as a way to define the CSS values used in components.
This change results in the .nav-link class not applying the correct
padding anymore to the component, so this commit manually adds the
padding back to keep the sidebar usable.

Signed-off-by: Davide Briani <davide.briani@secomind.com>