Update deps
There was a new ash and ash_postgres minor release, which includes useful
features for development (namely: generators), so include that.

Note that we have to do a slight change to tests since due to
https://github.com/ash-project/ash/issues/1282 the location calculation is not
reused when querying the position, so the test should expect 2 calls to the
mock. This could possibly change in the future when it is optimized again, so
leave a TODO detailing this.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>