base_image: correctly scope upload to the BaseImageCollection bucket
The extracted base_image_collection_id was nil in the current implementation,
leading to the image being put in the wrong bucket directory. This makes the
storage module behaviour more explicit in its requirement (instead of having to
pass the whole BaseImage) and adds a test to verify that the
base_image_collection_id received by the storage module actually matches the one
from the Base Image Collection.

Close #586

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>