Add reverse references to UpdateCampaign
UpdateCampaign has a foreign key on BaseImage and UpdateChannel. Without the
connected has_many relationships on the target resources a generic error was
shown when trying to delete a BaseImage or UpdateChannel that was related with
an existing UpdateCampaign (see #588).
This shows a more specific error message to the user.
Also update Ash since the bug fixed in 44c0d09e306 prevented the
UpdateChannel part to work.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>