BaseImage: fix missing 'version' when uploading an image
Commit 0e4d17b refactored how the scope gets passed to the
BaseImages.Storage behaviour, but in doing so the base image's version
was not provided anymore and thus the upload logic failed for a missing
'version' field, which is used to determine the filename for the
uploaded file.

This change adds the base image's version between the required
parameters by the upload logic.

Close #600

Signed-off-by: Davide Briani <davide.briani@secomind.com>