Make the existing database structure converge with Ash
Change the existing database structure with a migration on one side, and change
Ash resources to be consistent with the database on the other side.
Then, generate the initial resource snapshot with the AshPostgres migration
generator, and throw away the migrations (since we already have a set of
migrations to arrive at our current state).
This enables the use of the AshPostgres migration generator.

Signed-off-by: Riccardo Binetti <riccardo.binetti@secomind.com>