UI: Display deployed app's initial status as stopped instead of error
When an app is first deployed on a device, while the device has not
yet signaled that the deployment is ready, the deployment has a `nil`
status.
It's probably more appealing for the user to see an assumed `STOPPED`
status for the deployment instead of an `ERROR` status.
Later work can iterate on this improvement and properly define the
initial stages where the deployment is first "sent" to the device, and
it later becomes "ready" to be started.

Signed-off-by: Davide Briani <davide.briani@secomind.com>