UI: refetch application deployments without making them disappear
This change modifies the GraphQL refetchable fragment so that it is
defined on the RootQueryType instead of on Device.
This way, when refetching the fragment, Relay doesn't have to resort to
the `node` query that doesn't support loading relationships and would
return an empty set for the device's deployments.

Close #735

Signed-off-by: Davide Briani <davide.briani@secomind.com>