Handle incoming Starting deployment status when already Started
There are cases where starting an already Started deployment may happen.
In this cases, a Starting status is received on the DeploymentEvent
Astarte interface but no Started status is later received on the
AvailableDeployments interface since the deployment was already
available in a Started status.
The Deployment is then stuck in the Starting status.

To overcome this scenario, this change avoids handling a Starting status
coming from Astarte triggers when the deployment is already in the
Started status.

For simmetry, the same is done for the transition between the Stopping
and Stopped statuses.

Signed-off-by: Davide Briani <davide.briani@secomind.com>