feat(containers): network deployment caching
Adds `Edgehog.Containers.Network.Deployment` resource representing the
current state of a netowkr on a device.

`Netowrk.Deployment` works like the `Deployment` resource, representing
the `many_to_many` relationship between an image and a device. It keeps
track of the current state of a network on a device and gets updated
when the device publishes on `AvailableNetworks` interface.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>