feat(containers): container deployment caching
Adds `Edgehog.Containers.Container.Deployment` resource representing the
current state of the container on a device

`Container.Deployment` works like the `Deployment` resource, representing
the `many_to_many` relationship between an container and a device. It keeps
track of the current state of an container on the device and gets updated
when the device publishes on the `AvailableContainers` interface.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>