Document how to use Admin API to create tenants
Add documentation in the "Deploying with Kubernetes" guide on how to
create a new Edgehog tenant.
The relative section about initialization in the document is reworked to
propose using the Admin API as the preferred method for creating
tenants.
The manual method of issuing commands via an iEx session is kept
documented as a backup strategy.

Signed-off-by: Davide Briani <davide.briani@secomind.com>