Fix Device relationship not loaded in Campaign executor
Event handlers in the Update Campaign executor try to access the
`device` relationship either on the update target or the OTA operation
struct.
However, sometimes the relationship is not loaded beforehand and result
into a crash when handling events.
This change fixes the issue by reading the `.device_id` field directly,
which is defined by the belongs_to directive in the schemas of the
update target and OTA operation, without needing to access
`.device.device_id`.

Close #828

Signed-off-by: Davide Briani <davide.briani@secomind.com>