feat(containers): Add searchable app/release selection in deployments
Allow the user to search for a specific app when the user needs to
select an app (e.g. for deployment).
When trying to deploy or upgrade an app, the user needs to select the
app and the release (i.e. version). Instead of using a static dropdown,
the UI offers a search box to quickly filter and refine the options
shown in the dropdown.

Signed-off-by: Omar <omar.brbutovic@secomind.com>