fix(containers): notify campaign executor on deployment events
The executor for app deployment campaigns relies on PubSub notifications
to know how deployments are progressing.
However notifications were only published relative to the progress of
the deployment resources.
This change ensures that a notification is also published when the
deployment state is updated given the state reported by the state: e.g.
the deployment encountered an issue and enters an error state.
The campaign executor is now able to react to deployment events such as
errors.

Signed-off-by: Davide Briani <davide.briani@secomind.com>