fix(containers): use deployment's resources_state to check readiness
The app deployment campaigns executor relies on the `:ready?` field of
Deployment resources to know whether an app deployment was successful.
However, technically this field just exposes the fact that the
Deployment resource is advertised as acknowledged by the device, but it
does not mean that also all of its underlying resources are already
acknowledged by the device.

This change makes the executor use the `resources_state` field of the
Deployment resources to check for successful and complete rollout, since
this field becomes ready when all of its resources are ready (including
the deployment itself).

Signed-off-by: Davide Briani <davide.briani@secomind.com>