fix(containers): correctly compute deployable devices for a release
The previous calculation of deployable devices for a release did not
take into account that if the release has no requirements regarding
system models (i.e. system_models is an empty list), then it's
compatible with any device.
This change fixes the computation with a dedicated function to compute
whether system model requirements are satisfied, aiming for improved
readability.

Signed-off-by: Davide Briani <davide.briani@secomind.com>