chore(containers): remove default_preloads_for_target/0 function
The function is only called once, and it does more than it's needed, so
this change removes the function and just loads the needed associations
in place.
In addition, defining "default" preloads does not seem a great pattern,
as we might tend to use the function blindly regardless of the actual
needs of the specific logic we are implementing.

Signed-off-by: Davide Briani <davide.briani@secomind.com>