chore(docker-compose): add default fake key pairs for local development
Create and commit key pairs that can be used by default when running
Edgehog locally.
For instance, they can be used with the Docker Compose environment and
when running the seeding routine to create a sample Edgehog tenant for
development and testing.
Given that the key pairs are fixed, a set of pre-generated JWTs is also
available for testing, without having to manually generate them.

Signed-off-by: Davide Briani <davide.briani@secomind.com>