fix(containers): apply deployment campaign fixes (#914)
The executor for app deployment campaigns relies on PubSub notifications
to know how deployments are progressing.
However notifications were only published relative to the progress of
the deployment resources.
This change ensures that a notification is also published when the
deployment state is updated given the state reported by the state: e.g.
the deployment encountered an issue and enters an error state.
The campaign executor is now able to react to deployment events such as
errors.

The app deployment campaigns executor relies on the `:ready?` field of
Deployment resources to know whether an app deployment was successful.
However, technically this field just exposes the fact that the
Deployment resource is advertised as acknowledged by the device, but it
does not mean that also all of its underlying resources are already
acknowledged by the device.

This change makes the executor use the `resources_state` field of the
Deployment resources to check for successful and complete rollout, since
this field becomes ready when all of its resources are ready (including
the deployment itself).

The previous calculation of deployable devices for a release did not
take into account that if the release has no requirements regarding
system models (i.e. system_models is an empty list), then it's
compatible with any device.
This change fixes the computation with a dedicated function to compute
whether system model requirements are satisfied, aiming for improved
readability.

Use correct naming on references that have a wrong one due to copy &
paste mistakes.

The function is only called once, and it does more than it's needed, so
this change removes the function and just loads the needed associations
in place.
In addition, defining "default" preloads does not seem a great pattern,
as we might tend to use the function blindly regardless of the actual
needs of the specific logic we are implementing.

Signed-off-by: Davide Briani <davide.briani@secomind.com>