chore(justfile): Use Just task runner to quickly setup local dev env
Add a justfile that can be used with the Just task runner to quickly
setup a local Astarte + Edgehog setup, with a default tenant, and
connect a simulated device by using the Edgehog Device Runtime.
The setup currently relies on Docker Compose for Astarte and Edgehog,
and a local Rust installation to compile and run the device runtime.

Signed-off-by: Davide Briani <davide.briani@secomind.com>