chore(tooling): configure development tooling to quickly setup an Edgehog instance and connect a simulated device (#922)
Add a pgAdmin service to docker-compose.yml to facilitate the
development workflow.
The service allows to hit http://pgadmin.edgehog.localhost on the
browser to inspect and play with Edgehog database's contents.

Create and commit key pairs that can be used by default when running
Edgehog locally.
For instance, they can be used with the Docker Compose environment and
when running the seeding routine to create a sample Edgehog tenant for
development and testing.
Given that the key pairs are fixed, a set of pre-generated JWTs is also
available for testing, without having to manually generate them.

Since we are using Docker Compose for local development and testing, and
not for production scenarios, it makes sense to have the default
environment contain all relevant tooling needed to simulate and test
Edgehog functionalities.

Add a justfile that can be used with the Just task runner to quickly
setup a local Astarte + Edgehog setup, with a default tenant, and
connect a simulated device by using the Edgehog Device Runtime.
The setup currently relies on Docker Compose for Astarte and Edgehog,
and a local Rust installation to compile and run the device runtime.

Signed-off-by: Davide Briani <davide.briani@secomind.com>