fix: ensure migrations can be successfully rolled back
When migrating to the Ash framework, we gained the benefit of using
ash_postgres to automatically generate database migrations and resource
snasphots.
However the resulting migrations do not always run the needed operations
in the correct order, for the rollback routine.
This change fixes the needed migrations to ensure that operations are
done in the correct order and the database can be rolled back.

Signed-off-by: Davide Briani <davide.briani@secomind.com>