test: fix fixtures
Making a deployment fixture does not really look at the release it should
deploy. Insted it just creates a `Deployment` resource with the correct release.
This might however be a curse when trying to look at related resources.

`deployment_fixture` should instead create a valid deployment for the release,
which includes relationships.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>