fix(containers): refresh application list after deletion
Edgehog correctly deletes the application (when possible) but client-side
data isn't updated correctly. This is due to pagination and infinite
scrolling that were introduced after the delete functionality.
The delete functionality is now modified to use ConnectionHandler for
updating the relay store, so the deleted app is not visible anymore from
the UI.

Signed-off-by: Emina Muminovic <emina.muminovic@secomind.com>