Change the default value for CPU and memory settings from -1 to nil
Negative values are currently allowed for some fields in the backend
because the Astarte interfaces were originally defined this way the
Device Runtime interprets negative values as “not set” instead of
using null.

However, this represents internal and unintended complexity.
It should be isolated as much as possible, since it is not the expected
behavior for users managing container configurations.

Signed-off-by: Jasmina <jasmina.piric@secomind.com>