chore: update frontend to use deployment's isReady vs resourcesState
The resources_state field of deployments was removed in the backend as
the logic to verify the status of the deployment rollout was changed.
The field also gave the impression that resources were rolled out in
sequential steps, while instead Edgehog can receive out-of-order
acknowledgements; the API now expose the single state of each resource,
for API clients that want to inspect the rollout progress in details.
However, a single isReady boolean field is also exposed now to know
whether the resources are rolled out or not, for API clients that are
just interested in the overall deployment state.

The frontend is refactored to avoid displaying the ResourcesState.
It now just reports whether the deployment is still being rolled-out or
not.

Signed-off-by: Davide Briani <davide.briani@secomind.com>