docs: fix missing file error
The `pages/integrating/astarte_interfaces.md` file is automatically generated
in CI, so it was never added to the repo.
However, this requires contributors to either generate it, or adding a stopgap
empty file, and then remember not to commit the changes. Adding an empty
version will smooth out this process, not requiring any action by the
contributor.

Signed-off-by: Damiano Mason <damiano.mason@secomind.com>