ci: Use admin token to trigger CI workflows on ff (#1119)
A known *feature* of github actions does not allow CI to run after a fast
forward has been issued on a release branch. This might be a curse to
keep a linear history AND have images built on docker hub. By using an
`ADMIN_TOKEN` secret the action would run as an admin performed an
operation, making CI run after a fast forward.

Signed-off-by: Luca Zaninotto <luca.zaninotto@secomind.com>